backend/requirements.txt
# Django y DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.0

# Variables de entorno
python-decouple==3.8

# Validación y seguridad
django-filter==23.5
django-ratelimit==4.1.0

# Mercado Pago
mercadopago==2.2.1

# Cloudflare
requests==2.31.0
PyJWT==2.8.0

# Utilidades
python-dateutil==2.8.2
Pillow==10.1.0

# Producción
gunicorn==21.2.0
whitenoise==6.6.0
